Refactor commands to own file.
