+++ src/mr/developer/tests/test_commands.py
@@ -30,63 +30,63 @@ class TestCommand(TestCase):
         self.develop.auto_checkout = set(['foo', 'ham'])
 
     def testEmptyMatchList(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages([])
         self.assertEquals(pkgs, set(['foo', 'bar', 'baz', 'ham']))
 
     def testEmptyMatchListAuto(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages([], auto_checkout=True)
         self.assertEquals(pkgs, set(['foo', 'ham']))
 
     def testSingleArgMatchingOne(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ha'])
         self.assertEquals(pkgs, set(['ham']))
 
     def testSingleArgMatchingMultiple(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ba'])
         self.assertEquals(pkgs, set(['bar', 'baz']))
 
     def testArgsMatchingOne(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ha', 'zap'])
         self.assertEquals(pkgs, set(['ham']))
 
     def testArgsMatchingMultiple(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ba', 'zap'])
         self.assertEquals(pkgs, set(['bar', 'baz']))
 
     def testArgsMatchingMultiple2(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ha', 'ba'])
         self.assertEquals(pkgs, set(['bar', 'baz', 'ham']))
 
     def testSingleArgMatchingOneAuto(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ha'], auto_checkout=True)
         self.assertEquals(pkgs, set(['ham']))
 
     def testSingleArgMatchingMultipleAuto(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         self.assertRaises(SystemExit, Command(self.develop).get_packages,
                           ['ba'], auto_checkout=True)
 
     def testArgsMatchingOneAuto(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ha', 'zap'],
                                                   auto_checkout=True)
         self.assertEquals(pkgs, set(['ham']))
 
     def testArgsMatchingMultipleAuto(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         self.assertRaises(SystemExit, Command(self.develop).get_packages,
                           ['ba', 'zap'], auto_checkout=True)
 
     def testArgsMatchingMultiple2Auto(self):
-        from mr.developer.develop import Command
+        from mr.developer.commands import Command
         pkgs = Command(self.develop).get_packages(['ha', 'ba'],
                                                   auto_checkout=True)
         self.assertEquals(pkgs, set(['ham']))
@@ -94,7 +94,7 @@ class TestCommand(TestCase):
 
 class TestDeactivateCommand(TestCase):
     def setUp(self):
-        from mr.developer.develop import CmdDeactivate
+        from mr.developer.commands import CmdDeactivate
         self.develop = MockDevelop()
         self.develop.sources = dict(
             foo=MockSource(),
@@ -124,7 +124,7 @@ class TestDeactivateCommand(TestCase):
     def testDeactivateActivatedPackage(self):
         self.develop.config.develop['bar'] = True
         args = self.develop.parser.parse_args(args=['deactivate', 'bar'])
-        _logger = patch('mr.developer.develop.logger')
+        _logger = patch('mr.developer.commands.logger')
         logger = _logger.__enter__()
         try:
             self.cmd(args)
@@ -140,7 +140,7 @@ class TestDeactivateCommand(TestCase):
 
     def testDeactivateAutoCheckoutPackage(self):
         args = self.develop.parser.parse_args(args=['deactivate', 'foo'])
-        _logger = patch('mr.developer.develop.logger')
+        _logger = patch('mr.developer.commands.logger')
         logger = _logger.__enter__()
         try:
             self.cmd(args)
